requirements.txt
google-genai>=1.0.0
pyinstaller==6.11.1
